releases/v1.2.0/phobetron-v1.2.0/backend/requirements.txt
# Core Dependencies
alembic>=1.13.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
# geoalchemy2>=0.14.2  # Removed - PostGIS not available on Railway PostgreSQL
python-dotenv>=1.0.0

# Testing Dependencies
pytest>=8.0.0
pytest-postgresql>=5.0.0
pytest-alembic>=0.11.0

# Web Framework (FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Machine Learning & Data Science
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0

# Phase 2: Deep Learning (LSTM)
tensorflow>=2.15.0  # or tensorflow-cpu for non-GPU systems
keras>=2.15.0
joblib>=1.3.2  # Model serialization

# Phase 2: NLP & Sentiment Analysis
textblob>=0.17.1
nltk>=3.8.1
# python-hebrew>=0.1.0  # Optional: enhanced Hebrew support (not available on PyPI)
spacy>=3.7.0  # Advanced NLP

# Phase 2: External API Integration
newsapi-python>=0.2.7  # NewsAPI wrapper
tweepy>=4.14.0  # Twitter/X API v2
requests>=2.31.0
httpx>=0.27.0  # For async requests

# Additional Utilities
python-dateutil>=2.8.2
pytz>=2024.1
